refactor: Applied package hierarchy changes
