+++ src/app/src/main/java/com/imeanttobe/app901/ui/cart/components/CartItemCard.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.imeanttobe.app901.type.Cart
+import com.imeanttobe.app901.type.data.Cart
 import com.imeanttobe.app901.type.Formatter
 
 @Composable
